fix: smart-table interfaces exports
